Use safer migrations when deploying from scratch
The motivation for these changes is due to repeated errors reported by
cf teams attemping to deploy cf-networking from scratch. The 'legacy'
migrations perform multiple statements in one migration. Because mysql
is not transactional, when a migration timesout/fails, it is left in a
partially completed state. This changeset breaks up the original
migrations into individual migrations with one statement only.

[#159291931]

Signed-off-by: Aidan Obley <aobley@pivotal.io>