GetSubjectSpaces iterates paginated responses
The CC api would paginate the responses for multiple resources
including the spaces that a user belongs to. The default
pagination size is 50 elements.

If a user belongs to more than 50 spaces, the verification of
policies that the user has access to might fail, as not all the
spaces are retrieved properly.

This commit fixes the specific method cc_client.GetSubjectSpaces,
but a better approach would be create a PaginatedJSONClient
that would wrap json_client.JsonClient to iterate any
paginated response, but that might require more changes across
the code.