Update post-start script to handle go1.12 behavior
In go 1.12, misdirected http requests are handled differently than in
previous versions (https://golang.org/doc/go1.12#net/http). This causes
curl to return with an exit code of 22 now because of the 400 Bad Request being
returned by the golang HTTP server.

Since this is different behavior, the post-start script use to fail
to report that the server started successfully.

[#165295898](https://www.pivotaltracker.com/story/show/165295898)

Signed-off-by: Barack Karavani <bkaravani@pivotal.io>