Hacky race condition fix
"Enforce" order of curl commands by using a sleep. We need to associate
the user with an org before it can be associated with the spaces in that
org. For the time being, we're adding a sleep to attempt enforce this
order.

This hack _should_ be fixed with [this chore](https://www.pivotaltracker.com/story/show/166479329)

Co-authored-by: Barack Karavani <bkaravani@pivotal.io>
Co-authored-by: Clay Kauzlaric <ckauzlaric@pivotal.io>