Support TLS NATS in service-discovery-controller code
We want cf-deployment to ship NATS in TLS-only mode. To do that we need
everything in cf-deployment to support TLS connections to NATS.

This commit gives service-discovery-controller the ability to connect to
NATS over TLS. Only the source code is changed; the boshrelease config
will be changed in a later commit.