Add pagination race-condition checks to cc_client.GetSecurityGroups (#127)
If SGs are deleted between pages, there is a high chance that the offset
indices will result in SGs being missed, since 5000 ASGs are returned in
the first page. If this happens, it can result in the missing SGs' rules
being removed from active containers, which is a bad thing.

To guard against this, we decrement the per_page count on every page, to
ensure that the last ASG in the previous page matches the Nth indexed
ASG of the current request.