Use standard tas-runtime-{postgres,mysql-...} imgs
We created the following images off of our standard
cloudfoundry/tas-runtime-build image:
- cloudfoundry/tas-runtime-mysql-5.7
- cloudfoundry/tas-runtime-mysql-8.0
- cloudfoundry/tas-runtime-postgres
This commit switches the cf-networking-release scripts to use them
locally and in CI.

This also switches the default db when one invokes scripts/docker-shell
to postgres, which used to be the default.

Also removes things related to cf-networking-postgres-ifb because all of
the IFB feature work, including tests, was merged upstream to
containernetworking/plugins in like 2018 and we don't even run tests
related to this anymore. That is to say, we don't need the ifb image and
haven't used it in earnest for years at this point.

[#182691815](https://www.pivotaltracker.com/story/show/182691815)