Enable proxy's `/dumprequest` endpoint to return headers
This commit adds a query param to the proxy's `/dumprequest` endpoint
that adds two debug headers, `X-Proxy-Settable-Debug-Header` and
`X-Proxy-Immutable-Debug-Header`,  and also copies the headers made to the
proxy and sends them back in the response.

This is useful for checking how gorouter and other proxies handle
forwarding and returning headers. I used this when I was working on
acceptance for:
- https://github.com/cloudfoundry/gorouter/pull/356/files
- https://github.com/cloudfoundry/routing-release/pull/331